[Karate-Tools] Update to JDK 21
