+++ code/pom.xml
@@ -728,9 +728,13 @@
               <goal>verify</goal>
             </goals>
             <configuration>
+              <!-- Sets the VM argument line used when integration tests are run. -->
               <argLine>${failsafeArgLine}</argLine>
+              <!-- JDK21: Enable instrumentation for Mockito -->
+              <argLine>@{argLine} -javaagent:${org.mockito:mockito-core:jar}</argLine>
               <!-- Skips integration tests based on -DskipITs -->
               <skipTests>${skipITs}</skipTests>
+              <!-- Configures the test classes to be included in the integration test phase. -->
               <classesDirectory>${project.build.outputDirectory}</classesDirectory>
             </configuration>
           </execution>
@@ -758,6 +762,8 @@
         <configuration>
           <!-- Sets the VM argument line used when unit tests are run. -->
           <argLine>${surefireArgLine}</argLine>
+          <!-- JDK21: Enable instrumentation for Mockito -->
+          <argLine>@{argLine} -javaagent:${org.mockito:mockito-core:jar}</argLine>
           <!-- Skips unit tests based on -DskipUTs -->
           <skipTests>${skipUTs}</skipTests>
           <!-- Excludes integration tests when unit tests are run. -->
